CNN Input image adaption and dl4j version update
* Use constant player -1/1 image for layer 0 CNN Input instead of empty
fields
* Go to version 1.0.0-beta7 from dl4j
* Use java11

+++ neural-net-game/src/main/java/ch/evolutionsoft/net/game/NeuralNetConstants.java
@@ -8,7 +8,7 @@ public final class NeuralNetConstants {
     // Hide constructor
   }
 
-  public static final Integer DEFAULT_SEED = 235711;
+  public static final long DEFAULT_SEED = 235711;
 
   public static final Random randomGenerator = new Random(DEFAULT_SEED);
